+++ CHANGES.md
@@ -25,6 +25,8 @@ Changes 1.7.0-prerelease
 
 * X.509: add lc_x509_policy_cert_subject_match API
 
+* Linux kernel: add support for rfc4106(gcm(aes))
+
 Changes 1.6.0
 * ASN.1: use stack for small generator for small use cases